dpll: netlink: fix pin-id-get/device-id-get behavior
If pin was given a different pin label then the one searched by the user
it wrongly returned the match, despite there was not a match. Exaple
use case: pin had only a board-label assigned, a user requested search
of a pin with the label, logic returned a match in two cases:
- when pin actually had board-label match
- when a pin had no board-label but a different label present (either
  package-label or panel-label), even if there was no match

Do not return a match for a label if that label is not present on the
pin being checked.

Signed-off-by: Arkadiusz Kubalewski <arkadiusz.kubalewski@intel.com>